Cambio minimo

+++ UD1/src/ejercicios/ej04.java
@@ -6,10 +6,10 @@ public class ej04 {
 
 	public static void main(String[] args) {
 		Scanner s = new Scanner(System.in);
-		double meses = 12, mes = 8;
+		int meses = 12, mes = 8;
 		double porcentaje;
 
-		porcentaje = mes * 100 / meses;
+		porcentaje = (double) mes * 100 / (double) meses;
 
 		System.out.println("Mes " + mes + " de " + meses);
 		System.out.printf("Porcentaje transcurrido: %.4f", porcentaje, "%");
